+++ BusinessLogic/Inventory.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Text;
 
 using BLToolkit.Data;
+using BLToolkit.Data.DataProvider;
 using BLToolkit.DataAccess;
 using BLToolkit.EditableObjects;
 
@@ -14,7 +15,7 @@ namespace ItemsUsage.BusinessLogic
   [TableName("inventories")]
   public abstract class Inventory : EditableObject<Inventory>
   {
-    [PrimaryKey]
+    [PrimaryKey, NonUpdatable]
     public abstract int     Id          { get; set; }
     public abstract string  Code        { get; set; }
     public abstract string  Description { get; set; }
@@ -33,6 +34,12 @@ namespace ItemsUsage.BusinessLogic
 
         return _query;
       }
+    }
+
+    public EditableList<Inventory> GetAll(DbManager db)
+    {
+      EditableList<Inventory> list = new EditableList<Inventory>();
+      return Query.SelectAll(db, list);
     }
 
     public Inventory Get(DbManager db, int id)
@@ -54,8 +61,19 @@ namespace ItemsUsage.BusinessLogic
       return item.Id > 0;
     }
 
+    public bool InUse(DbManager db, Inventory item)
+    {
+      string sql = "select count(*) from order_items where inventory_id = "
+        + db.DataProvider.Convert("InvId", ConvertType.NameToQueryParameter);
+
+      return db.SetCommand(sql, db.Parameter("InvId", item.Id)).ExecuteScalar<int>() > 0;
+    }
+
     public bool Delete(DbManager db, Inventory item)
     {
+      if (InUse(db, item))
+        throw new ApplicationException("This inventory is in use and can't be deleted.");
+      
       return Query.Delete(db, item) > 0;
     }
   }
